Python 2 --> Python 3 fixes
No python 3 fixes have actually been made yet.

Renaming/replacing WavePy with WavePy3

Adding utils

adding some schmidt ft functions

fixing test

fixing test for WavePy3

adding todo and changing references to wavepy to WavePy3

Added a doc string. This needs to be improved considerably.

Added the start of a docstring. This needs to be improved in order to
describe how to actually use the package. There is no point in doing
this yet since most of the functionality and the way to interface with
it is not complete.

Some minor syntax fixes too.